requirements.txt
# ============================================================================
# Oracle 23ai RAG Chatbot - Python Dependencies
# ============================================================================

# Core Dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
streamlit>=1.32.0

# Oracle Database
oracledb>=2.0.0

# OCI SDK
oci>=2.100.0

# LlamaIndex Core
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-embeddings-oci-genai>=0.1.0
llama-index-llms-oci-genai>=0.1.0

# Optional: Reranker
llama-index-postprocessor-cohere-rerank>=0.1.0

# Document Processing
pypdf>=4.0.0
tokenizers>=0.15.0

# Utilities
tqdm>=4.66.0
numpy>=1.26.0

# Optional: Observability
arize-phoenix>=3.0.0
llama-index-callbacks-arize-phoenix>=0.1.0
openinference-instrumentation-llama-index>=1.0.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-exporter-otlp-proto-http>=1.23.0
